% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pydpiperQC.R
\name{launch}
\alias{launch}
\title{Launch the pydpiperQC application in an external browser}
\usage{
launch(annotation = NULL, consensus = NULL, wd = getwd())
}
\arguments{
\item{annotation}{A path to an annotation file.}

\item{consensus}{A path to a consensus MINC file.}

\item{wd}{A path to the working directory.}
}
\description{
Launch the pydpiperQC application in an external browser
}
\seealso{
\code{\link{launch_example}}
}
